input: libinput: add setting to allow changing keymap layout
This is a bit of a hack but it's needed in order for us to allow
using a different keyboard layout. The idea in the future is to
move libinput to a peripheral add-on but that won't be until
V19. So this is needed in the meantime.

I can also add more options in the future related to mice and keyboards
but I have only heard about keyboard layouts so far.