Add new backlight interface and gui
this adds a new interface to be able to access a hardware
backlight. this also adds the gui elements needed to show
the backlight brightness change and to change it via the
touch screen interface. this change is quite invasive but
it is based off the way the volume control works.

the backlight controls will not be show if no backlight
has been register to CServiceBroker